Completed the logic for logging in, logging out, and checks for user role

+++ django_part_2/lazone/shop/views.py
@@ -1,6 +1,8 @@
 from django.shortcuts import render, redirect
+from django.contrib.auth.decorators import login_required
 from .models import Item
 from .forms import ItemForm
+from customers.models import StoreUser
 
 # Create your views here.
 def index(request):
@@ -16,32 +18,48 @@ def detail_view(request, item_id):
 
     return render(request, 'shop/detail.html', {'product': product})
 
+@login_required
 def create_product(request):
-    form = ItemForm(request.POST or None)
 
-    if form.is_valid():
-        form.save()
-        return redirect('shop:product_list_view')
-    
-    return render(request, 'shop/product_form.html', {'form':form})
+    customer_account = StoreUser.objects.get(user__pk = request.user.id)
+    if customer_account.role == 'admin':
+        form = ItemForm(request.POST or None)
 
+        if form.is_valid():
+            form.save()
+            return redirect('shop:product_list_view')
+        
+        return render(request, 'shop/product_form.html', {'form':form})
+    return redirect('customers:access_denied')
+
+@login_required
 def edit_product(request, item_id):
-    product = Item.objects.get(pk = item_id)
-    form = ItemForm(request.POST or None, instance = product)
 
-    if form.is_valid():
-        form.save()
-        return redirect('shop:product_list_view')
-    
-    return render(request, 'shop/product_form.html', {'form':form})
+    customer_account = StoreUser.objects.get(user__pk = request.user.id)
+    if customer_account.role == 'admin':
+        product = Item.objects.get(pk = item_id)
+        form = ItemForm(request.POST or None, instance = product)
 
+        if form.is_valid():
+            form.save()
+            return redirect('shop:product_list_view')
+        
+        return render(request, 'shop/product_form.html', {'form':form})
+    return redirect('customers:access_denied')
+
+@login_required
 def delete_product(request, item_id):
-    product = Item.objects.get(pk = item_id)
 
-    if request.method == 'POST':
-        product.delete()
-        return redirect('shop:product_list_view')
-    
-    return render(request, 'shop/delete.html', {'product':product})
+    customer_account = StoreUser.objects.get(user__pk = request.user.id)
+    if customer_account.role == 'admin':
+
+        product = Item.objects.get(pk = item_id)
+
+        if request.method == 'POST':
+            product.delete()
+            return redirect('shop:product_list_view')
+        
+        return render(request, 'shop/delete.html', {'product':product})
+    return redirect('customers:access_denied')
